Move the `range` fields to the module too.
