+++ modules/mapper-extras/src/main/java/org/elasticsearch/index/mapper/BinaryRangeUtil.java
