+++ modules/percolator/build.gradle
@@ -25,6 +25,7 @@ esplugin {
 
 dependencies {
     // for testing hasChild and hasParent rejections
+    compile project(path: ':modules:mapper-extras', configuration: 'runtime')
     testCompile project(path: ':modules:parent-join', configuration: 'runtime')
 }
 compileJava.options.compilerArgs << "-Xlint:-deprecation,-rawtypes"
